+++ lang/attribute/program/src/declare_program/mods/errors.rs
@@ -4,14 +4,16 @@ use quote::{format_ident, quote};
 pub fn gen_errors_mod(idl: &Idl) -> proc_macro2::TokenStream {
     let errors = idl.errors.iter().map(|e| {
         let name = format_ident!("{}", e.name);
+        let code = e.code;
         quote! {
-            #name,
+            #name = #code,
         }
     });
 
     if errors.len() == 0 {
         return quote! {
             /// Program error type definitions.
+            #[cfg(not(feature = "idl-build"))]
             pub mod errors {
             }
         };
@@ -19,9 +21,10 @@ pub fn gen_errors_mod(idl: &Idl) -> proc_macro2::TokenStream {
 
     quote! {
         /// Program error type definitions.
+        #[cfg(not(feature = "idl-build"))]
         pub mod errors {
 
-            #[anchor_lang::error_code]
+            #[anchor_lang::error_code(offset = 0)]
             pub enum ProgramError {
                 #(#errors)*
             }