+++ lang/attribute/program/src/declare_program/mods/internal.rs
@@ -1,6 +1,4 @@
-use anchor_lang_idl::types::{
-    Idl, IdlInstruction, IdlInstructionAccountItem, IdlInstructionAccounts,
-};
+use anchor_lang_idl::types::{Idl, IdlInstructionAccountItem};
 use anchor_syn::{
     codegen::accounts::{__client_accounts, __cpi_client_accounts},
     parser::accounts,
@@ -9,7 +7,10 @@ use anchor_syn::{
 use heck::CamelCase;
 use quote::{format_ident, quote};
 
-use super::common::{convert_idl_type_to_syn_type, gen_discriminator, get_canonical_program_id};
+use super::common::{
+    convert_idl_type_to_syn_type, gen_discriminator, get_all_instruction_accounts,
+    get_canonical_program_id,
+};
 
 pub fn gen_internal_mod(idl: &Idl) -> proc_macro2::TokenStream {
     let internal_args_mod = gen_internal_args_mod(idl);
@@ -107,62 +108,17 @@ fn gen_internal_accounts_common(
     idl: &Idl,
     gen_accounts: impl Fn(&AccountsStruct, proc_macro2::TokenStream) -> proc_macro2::TokenStream,
 ) -> proc_macro2::TokenStream {
-    // It's possible to declare an accounts struct and not use it as an instruction, see
-    // https://github.com/coral-xyz/anchor/issues/3274
-    fn get_non_instruction_composite_accounts<'a>(
-        accs: &'a [IdlInstructionAccountItem],
-        idl: &'a Idl,
-    ) -> Vec<&'a IdlInstructionAccounts> {
-        accs.iter()
-            .flat_map(|acc| match acc {
-                IdlInstructionAccountItem::Composite(accs)
-                    if !idl
-                        .instructions
-                        .iter()
-                        .any(|ix| ix.accounts == accs.accounts) =>
-                {
-                    let mut non_ix_composite_accs =
-                        get_non_instruction_composite_accounts(&accs.accounts, idl);
-                    if !non_ix_composite_accs.contains(&accs) {
-                        non_ix_composite_accs.push(accs);
-                    }
-                    non_ix_composite_accs
-                }
-                _ => Default::default(),
-            })
-            .collect()
-    }
-
-    let ix_accs = idl
-        .instructions
+    let all_ix_accs = get_all_instruction_accounts(idl);
+    let accounts = all_ix_accs
         .iter()
-        .flat_map(|ix| ix.accounts.to_owned())
-        .collect::<Vec<_>>();
-    let combined_ixs = get_non_instruction_composite_accounts(&ix_accs, idl)
-        .into_iter()
-        .map(|accs| IdlInstruction {
-            // The name is not guaranteed to be the same as the one used in the actual source code
-            // of the program because the IDL only stores the field names.
-            name: accs.name.to_owned(),
-            accounts: accs.accounts.to_owned(),
-            args: Default::default(),
-            discriminator: Default::default(),
-            docs: Default::default(),
-            returns: Default::default(),
-        })
-        .chain(idl.instructions.iter().cloned())
-        .collect::<Vec<_>>();
-
-    let accounts = combined_ixs
-        .iter()
-        .map(|ix| {
-            let ident = format_ident!("{}", ix.name.to_camel_case());
-            let generics = if ix.accounts.is_empty() {
+        .map(|accs| {
+            let ident = format_ident!("{}", accs.name.to_camel_case());
+            let generics = if accs.accounts.is_empty() {
                 quote!()
             } else {
                 quote!(<'info>)
             };
-            let accounts = ix.accounts.iter().map(|acc| match acc {
+            let accounts = accs.accounts.iter().map(|acc| match acc {
                 IdlInstructionAccountItem::Single(acc) => {
                     let name = format_ident!("{}", acc.name);
 
@@ -191,11 +147,11 @@ fn gen_internal_accounts_common(
                 }
                 IdlInstructionAccountItem::Composite(accs) => {
                     let name = format_ident!("{}", accs.name);
-                    let ty_name = combined_ixs
+                    let ty_name = all_ix_accs
                         .iter()
-                        .find(|ix| ix.accounts == accs.accounts)
-                        .map(|ix| format_ident!("{}", ix.name.to_camel_case()))
-                        .expect("Instruction must exist");
+                        .find(|a| a.accounts == accs.accounts)
+                        .map(|a| format_ident!("{}", a.name.to_camel_case()))
+                        .expect("Accounts must exist");
 
                     quote! {
                         pub #name: #ty_name #generics